Check for existence of  rl-deepracer-sagemaker before deleting!

+++ scripts/training/set-last-run-to-pretrained.sh
@@ -1,4 +1,12 @@
 #!/usr/bin/env bash
+Folder=rl-deepracer-sagemaker
+if [ -d ../../docker/volumes/minio/bucket/rl-deepracer-sagemaker ];
+then
+	echo "Folder $Folder  exist."
+	rm -rf ../../docker/volumes/minio/bucket/rl-deepracer-pretrained
+	mv ../../docker/volumes/minio/bucket/rl-deepracer-sagemaker ../../docker/volumes/minio/bucket/rl-deepracer-pretrained
+	echo "Done."
 
-rm -rf ../../docker/volumes/minio/bucket/rl-deepracer-pretrained
-mv ../../docker/volumes/minio/bucket/rl-deepracer-sagemaker ../../docker/volumes/minio/bucket/rl-deepracer-pretrained
+else
+	echo "Folder $Folder does not exist" 
+fi
